Oauth validation (#25)
* feature(user auth)

* fix(disable auth in properties for now)

* fix(remove user table; add version numbers in buiid)

* fix(enable auth in stg)

+++ src/main/groovy/com/temp/aggregation/kelvinapi/domain/ListResponse.groovy
@@ -2,5 +2,5 @@ package com.temp.aggregation.kelvinapi.domain
 
 class ListResponse<T> {
   List<T> results
-  int total
+  long total
 }
